requirements.txt
colorama==0.4.6
deep_translator==1.11.4
discord_py_interactions==5.11.0
meross_iot==0.4.6.0
Pillow==10.2.0
python-dotenv==1.0.1
webcolors==1.13
